Move tooltipOptions to config.ts, use for CDataTable
