+++ frontend/src/maker/components/CDataTable.vue
@@ -149,7 +149,8 @@ async function initDataTableWArray(data: KeyValueArray, isReplace = true) {
 
   const container = await embed('#map-vis', <VisualizationSpec>versionSpec, {
     renderer: 'svg',
-    actions: false
+    actions: false,
+    tooltip: config.tooltipOptions
   })
   visView = container.view
   if (props.mapColorScheme !== 'custom')
@@ -427,24 +428,6 @@ table input[type='color'] {
 .cell-content {
   position: relative;
   display: inline-block;
-}
-
-/* Tooltip styling */
-.tooltip {
-  visibility: hidden;
-  background-color: #333;
-  color: #fff;
-  text-align: left;
-  padding: 4px;
-  border-radius: 4px;
-  position: absolute;
-  z-index: 10;
-  left: 100%;
-  top: 50%;
-  transform: translateY(-50%);
-  white-space: nowrap;
-  opacity: 0;
-  transition: opacity 0s;
 }
 
 /* Show tooltip immediately on hover */
